split the card component into its own component
