--- frontend/src/components/timeTracker/timetracker.scss
@@ -1,19 +1,0 @@
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1050;
-
-  .card {
-    width: 95%;
-    height: 95%;
-    background: #fff; // or whatever you want
-    border-radius: 0; // removes rounded corners so it fits perfectly
-  }
-}
